+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 - [In Progress: 30 Jul 2021] Support `ContactPatientSheet` with no appointment
 
 ### Changes
+- Don't load search results if search query is not changed from previous search query
+- Show keyboard and prefill search query if present when instant search screen is shown
 - [In Progress: 22 Jul 2021] Add option to download & share overdue list
 - [In Progress: 14 Jul 2021] Medication screen improvements
   - Search for commonly used drugs
@@ -73,7 +75,6 @@
 - Increase the search and scan button height in patients screen
 - Show new video illustration in patients screen
 - After language is changed app will go back to home screen
-- Don't load search results if search query is not changed from previous search query
   
 ### Fixes
 - Fix `ContactPatientBottomSheet` UI spacing and styling